web-app/backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework==3.14.0
mysql-client==0.0.1
mysqlclient
django-cors-headers==3.14.0
requests